fix(auth): verify passkey assertion in-process instead of fetching own URL
The NextAuth passkey provider verified assertions by fetching
`${NEXTAUTH_URL}/api/passkey/authenticate/verify-authentication` — an HTTP
call from the server back to its own public hostname. On the new Coolify/Vultr
infra there's no hairpin-NAT loopback, so the connection timed out
(ConnectTimeoutError on volunteers.everybodyeats.nz:443) and every passkey
login failed.

Extract the verification ceremony into a shared in-process function
(`verifyPasskeyAuthentication` in src/lib/passkey-auth.ts) and call it directly
from both the NextAuth `authorize()` callback and the HTTP route, which is now a
thin wrapper. No network hop, no dependency on the server reaching its own
public URL.
